Sensors/Imagenex837B: Revert modification to fix bug.

+++ src/Sensors/Imagenex837B/Frame.hpp
@@ -187,7 +187,7 @@ namespace Sensors
       {
         roll = Angles::degrees(roll);
 
-        ByteCopy::toBE(((uint16_t)(roll * 10 + 900) & 0x7fff) | 0x8000,
+        ByteCopy::toBE((uint16_t)(((uint16_t)(roll * 10 + 900) & 0x7fff) | 0x8000),
                        getData() + getIndexRoll());
 
         if (getFooterData() != NULL)
@@ -202,7 +202,7 @@ namespace Sensors
       {
         pitch = Angles::degrees(pitch);
 
-        ByteCopy::toBE(((uint16_t)(pitch * 10 + 900) & 0x7fff) | 0x8000,
+        ByteCopy::toBE((uint16_t)(((uint16_t)(pitch * 10 + 900) & 0x7fff) | 0x8000),
                        getData() + getIndexPitch());
 
         if (getFooterData() != NULL)
@@ -220,7 +220,7 @@ namespace Sensors
 
         heading = Angles::degrees(heading);
 
-        ByteCopy::toBE(((uint16_t)(heading * 10) & 0x7fff) | 0x8000,
+        ByteCopy::toBE((uint16_t)(((uint16_t)(heading * 10) & 0x7fff) | 0x8000),
                        getData() + getIndexHeading());
 
         if (getFooterData() != NULL)
